pyproject.toml
[tool.poetry]
name = "se_gym"
version = "0.0.2"
authors = [
    "Dr. Kyrill Schmid <kyrillschmid@web.de>",
    "Gerhard Stenzel <gerhard.stenzel@ifi.lmu.de>"
]
readme = "README.md"
description = "SEGym allows you to apply patches to Python packages and run tests in isolated environments."
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = ">=3.9.1,<3.9.7 || >3.9.7,<3.13"
annotated-types = "^0.6.0"
docker = "^7.0.0"
iniconfig = "^2.0.0"
openai = "1.26.0"
pydantic = "^2.6.4"
pyproject_hooks = "^1.0.0"
pytest = "^8.1.1"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
tqdm = "^4.66.2"
typing_extensions = "^4.11.0"
instructor = "1.2.5"
tenacity = "8.3.0"
datasets = "^2.19.1"
rank_bm25 = "^0.2.2"
regex = "^2024.5.15"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.25.0"
pyarrow = "^16.0.0"
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
exclude = ["./temp/**"]
